requirements.txt
google-cloud-bigquery
openai
flask
requests
python-dotenv
gunicorn